strands-agent-streamlit/docker_app/requirements.txt
streamlit==1.45.1
boto3==1.38.18
streamlit-cognito-auth==1.3.1
strands-agents>=1.15.0
strands-agents-tools>=0.1.1